+++ react/components/Link.tsx
@@ -1,23 +1,8 @@
-import React, {
-  MouseEvent,
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-} from 'react'
-import { parse } from 'graphql'
+import React, { MouseEvent, useCallback, useMemo } from 'react'
 import { useInView } from 'react-intersection-observer'
-import { useApolloClient } from 'react-apollo'
-import {
-  NavigateOptions,
-  pathFromPageName,
-  getNavigationRouteToNavigate,
-  queryStringToMap,
-} from '../utils/pages'
+import { NavigateOptions, pathFromPageName } from '../utils/pages'
 import { useRuntime } from './RenderContext'
-import { getPrefetchForPath, fetchRouteData } from '../utils/routes'
-import { usePrefetch, PrefetchState } from './Prefetch/PrefetchContext'
-import { fetchComponents } from '../utils/components'
+import { usePrefetchAttempt } from '../hooks/prefetch'
 
 const isLeftClickEvent = (event: MouseEvent<HTMLAnchorElement>) =>
   event.button === 0
@@ -39,205 +24,6 @@ interface Props extends NavigateOptions {
   target?: string
 }
 
-const getPageToNavigate = (path: string) => {
-  const query = queryStringToMap(location.search) as RenderRuntime['query']
-  return getPrefetchForPath({
-    path: path,
-    fetcher: fetch,
-    query,
-  })
-}
-
-const hydrateApolloCache = (
-  queryData: Array<{
-    data: string
-    query: string
-    variables: Record<string, any>
-  }>,
-  client: ApolloClientType
-) => {
-  return queryData.map(({ data, query, variables }) => {
-    try {
-      client.writeQuery({
-        query: parse(query),
-        data: JSON.parse(data),
-        variables,
-      })
-    } catch (e) {
-      // do nothing
-    }
-  })
-}
-
-interface PrefetchRequestsArgs {
-  client: ApolloClientType
-  navigationRoute: any
-  page?: string
-  pages: RenderRuntime['pages']
-  prefetchState: PrefetchState
-  hints: RenderRuntime['hints']
-  renderMajor: RenderRuntime['renderMajor']
-}
-
-const prefetchRequests = async ({
-  client,
-  navigationRoute,
-  page,
-  pages,
-  prefetchState,
-  hints,
-  renderMajor,
-}: PrefetchRequestsArgs) => {
-  const { pathsCache, pathsState, routesCache, routePromise } = prefetchState
-  const navigationData = await getPageToNavigate(navigationRoute.path)
-
-  const navigationPage = page ?? navigationData.page
-
-  pathsState[navigationRoute.path] = { fetching: false, page: navigationPage }
-
-  const declarer = pages[navigationPage]?.declarer
-  if (!declarer) {
-    return
-  }
-
-  if (navigationData.queryData) {
-    hydrateApolloCache(navigationData.queryData, client)
-  }
-
-  const cacheObj = {
-    routeId: navigationPage,
-    matchingPage: navigationData.route,
-    contentResponse: navigationData.contentResponse,
-  }
-
-  if (navigationPage === 'store.product') {
-    pathsCache.product.set(navigationRoute.path, cacheObj)
-  } else if (navigationPage.startsWith('store.search')) {
-    pathsCache.search.set(navigationRoute.path, cacheObj)
-  } else {
-    pathsCache.other.set(navigationRoute.path, cacheObj)
-  }
-
-  let routeDataCache = routesCache.get(navigationPage)
-  const routePromiseValue = routePromise[navigationPage]
-
-  if (!routeDataCache && !routePromiseValue) {
-    const newState = {
-      promisePending: true,
-      promise: fetchRouteData({
-        apolloClient: client,
-        routeId: navigationPage,
-        declarer,
-        hints,
-        renderMajor,
-      }),
-    }
-    newState.promise
-      .then((data) => {
-        routesCache.set(navigationPage, data)
-      })
-      .finally(() => {
-        newState.promisePending = false
-        newState.promise = null as any
-        routePromise[navigationPage] = null
-      })
-    // newState.promise.then(data => {
-    //   newState.promisePending = false
-    //   routesCache.set(navigationPage, data)
-    //   newState.promise = null
-    //   routePromise[navigationPage] = null
-    // }).catch(() => {
-    //   newState.promisePending = false
-    //   newState.promise = null
-    //   routePromise[navigationPage] = null
-    // })
-
-    routePromise[navigationPage] = newState
-  }
-  const globalRouteDataState = routePromise[navigationPage]
-  if (globalRouteDataState?.promisePending) {
-    await globalRouteDataState.promise
-    routeDataCache = routesCache.get(navigationPage)
-  }
-
-  if (!routeDataCache) {
-    return
-  }
-
-  await fetchComponents(
-    routeDataCache.components,
-    window.__RUNTIME__,
-    routeDataCache.extensions
-  )
-}
-
-interface UsePrefetchArgs {
-  inView: boolean
-  page?: string
-  href: string
-  options: NavigateOptions
-}
-
-const usePrefetchAttempt = ({
-  inView,
-  page,
-  href,
-  options,
-}: UsePrefetchArgs) => {
-  const runtime = useRuntime()
-  const prefetchState = usePrefetch()
-  const client = useApolloClient() as ApolloClientType
-  const hasTried = useRef(false)
-
-  const { pages, navigationRouteModifiers, hints, renderMajor } = runtime
-
-  useEffect(() => {
-    if (inView && !hasTried.current) {
-      const { pathsState, queue } = prefetchState
-      if (href && href[0] !== '/') {
-        // so funciona com path relativo
-        return
-      }
-
-      options.modifiers = navigationRouteModifiers
-      const navigationRoute = getNavigationRouteToNavigate(pages, options)
-
-      if (pathsState[navigationRoute.path]) {
-        return //dont redo
-      }
-      pathsState[navigationRoute.path] = { fetching: true }
-
-      const priority = page === 'store.product' ? 1 : 0
-      hasTried.current = true
-
-      queue.add(
-        async () =>
-          prefetchRequests({
-            client,
-            navigationRoute,
-            page,
-            pages,
-            prefetchState,
-            hints,
-            renderMajor,
-          }),
-        { priority }
-      )
-    }
-  }, [
-    client,
-    hints,
-    href,
-    inView,
-    navigationRouteModifiers,
-    options,
-    page,
-    pages,
-    prefetchState,
-    renderMajor,
-  ])
-}
-
 const Link: React.FunctionComponent<Props> = ({
   page,
   onClick = () => {},
@@ -257,7 +43,6 @@ const Link: React.FunctionComponent<Props> = ({
     navigate,
     rootPath = '',
     route: { domain },
-    navigationRouteModifiers,
   } = useRuntime()
 
   const options = useMemo(
