--- react/queries/pageContent.graphql
@@ -1,13 +1,0 @@
-query PageContent($input: PageContentInput) {
-  pageContent(input: $input) {
-    contentMapJSON
-    userMessages {
-      key
-      message {
-        from
-        context
-        content
-      }
-    }
-  }
-}
