+++ react/typings/global.d.ts
@@ -439,14 +439,6 @@ declare global {
     canonicalBaseAddress: string
   }
 
-  // interface DeviceHints {
-  //   desktop: boolean
-  //   mobile: boolean
-  //   phone: boolean
-  //   tablet: boolean
-  //   unknown: boolean
-  // }
-
   interface RenderRuntime {
     amp: boolean
     account: string
@@ -499,7 +491,6 @@ declare global {
       variables: any
       data: string
     }>
-    fidelis: any
   }
 
   interface CacheHints {
