+++ react/utils/pages.ts
@@ -335,69 +335,6 @@ export function navigate(
     return false
   }
 
-  // if (!page && !inputTo) {
-  //   console.error(
-  //     `Invalid navigation options. You should use 'page' or 'to' parameters`
-  //   )
-  //   return false
-  // }
-
-  // if (inputTo && inputQuery) {
-  //   console.warn(
-  //     `You shouldn't pass 'query' in a separate prop when using 'to'`
-  //   )
-  // }
-
-  // If the prop `to` is something like `to="#header"`
-  // just change the hash using location, avoid doing a history navigation
-  // if (inputTo.indexOf('#') === 0 && inputTo.indexOf('?') === -1) {
-  //   window.location.hash = inputTo
-  //   return true
-  // }
-
-  // const [to, extractedQuery] = (is(String, inputTo) ? inputTo : '').split('?')
-  // const [realQuery, hash] = (is(String, extractedQuery)
-  //   ? extractedQuery
-  //   : ''
-  // ).split('#')
-  // const realHash = is(String, hash) ? `#${hash}` : ''
-  // let query = inputQuery || realQuery
-
-  // let navigationRoute: any = {}
-
-  // if (isEnabled('RENDER_NAVIGATION')) {
-  //   const fallbackPage = { path: to, params: {}, id: '' }
-  //   const routeFromPage = page && getRouteFromPageName(page, pages, params)
-  //   const routeFromPath = getRouteFromPath(to, pages)
-  //   navigationRoute = routeFromPage || routeFromPath || fallbackPage
-  // } else {
-  //   navigationRoute = page
-  //     ? getRouteFromPageName(page, pages, params)
-  //     : getRouteFromPathOld(to, pages, query, realHash)
-  // }
-
-  // if (!navigationRoute) {
-  //   console.warn(
-  //     `Unable to find route for ${
-  //     page ? `page '${page}' and the passed parameters` : `path '${to}'`
-  //     }`
-  //   )
-  //   return false
-  // }
-
-  // navigationRoute.path = navigationRootPath(navigationRoute.path, rootPath)
-  // if (modifiers) {
-  //   for (const modifier of modifiers) {
-  //     const { path, query: fixedQuery } = modifier({
-  //       path: navigationRoute.path,
-  //       query,
-  //       options: modifiersOptions,
-  //     })
-  //     navigationRoute.path = path || navigationRoute.path
-  //     query = fixedQuery || query
-  //   }
-  // }
-
   if (history) {
     const nextQuery = mergePersistingQueries(
       history.location.search,
